feat: get chain info using chain id
